add placeUsecases files for atomic db operations
