+++ src/usecases/placeUsecases/createSinglePlace.js
@@ -1,0 +1,3 @@
+const createSinglePlace = async (place) => await Place.create(place)
+
+export { createSinglePlace }
