+++ src/usecases/placeUsecases/deleteSinglePlace.js
@@ -1,0 +1,3 @@
+const deleteSinglePlace = async (id) => await Place.findByIdAndDelete(id)
+
+export { deleteSinglePlace }
