+++ src/usecases/placeUsecases/getAllPlaces.js
@@ -1,0 +1,3 @@
+const getAllPlaces = async () => await Place.find()
+
+export { getAllPlaces }
