+++ src/usecases/placeUsecases/getSinglePlace.js
@@ -1,0 +1,3 @@
+const getSinglePlace = async (id) => await Place.findById(id)
+
+export { getSinglePlace }
