--- src/usecases/placeUsecases/place.usecase.js
@@ -1,20 +1,0 @@
-import { Place } from '../models/place.model.js'
-
-const getAllPlaces = async () => await Place.find()
-
-const createSinglePlace = async (place) => await Place.create(place)
-
-const getSinglePlace = async (id) => await Place.findById(id)
-
-const updateSinglePlace = async (id, body) =>
-  await Place.findByIdAndUpdate(id, body)
-
-const deleteSinglePlace = async (id) => await Place.findByIdAndDelete(id)
-
-export {
-  getAllPlaces,
-  createSinglePlace,
-  getSinglePlace,
-  updateSinglePlace,
-  deleteSinglePlace,
-}
