+++ src/usecases/placeUsecases/updateSinglePlace.js
@@ -1,0 +1,4 @@
+const updateSinglePlace = async (id, body) =>
+  await Place.findByIdAndUpdate(id, body)
+
+export { updateSinglePlace }
