+++ src/usecases/routeUsecases/route.usecase.js
@@ -1,4 +1,4 @@
-import { Route } from '../models/route.model.js'
+import { Route } from '../../models/route.model.js'
 
 const getAllRoutes = async () => await Route.find()
